Changes added to use the extended methods available in jwt extended like blacklisting refreshing token, logging out users, etc.

+++ resources/item.py
@@ -1,6 +1,5 @@
-import sqlite3
 from flask_restful import Resource, reqparse
-from flask_jwt_extended import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt_claims, jwt_optional, get_jwt_identity, fresh_jwt_required
 from models.item import ItemModel
 
 items = []
@@ -24,6 +23,7 @@ class Item(Resource):
             return item.json()
         return {"message": "Item {} not found".format(name)}, 404
 
+    @fresh_jwt_required
     def post(self, name):
         if ItemModel.find_by_name(name):
             return {'message': "Item {} already exists".format(name)}, 400  # HTTP status code 400 for bad request
@@ -38,12 +38,17 @@ class Item(Resource):
 
         return item.json(), 201   # HTTP status code 201 when some object is created
 
+    @jwt_required
     def delete(self, name):
+        claims = get_jwt_claims()
+        if not claims['is_admin']:
+            return {"message": "Admin previliges required"}, 401
+
         item = ItemModel.find_by_name(name)
         if item:
             item.delete_from_db()
-
-        return {'message': 'item deleted'}
+            return {'message': 'item deleted'}
+        return {'message': 'Item not found'}, 404
 
     def put(self, name):
         data = Item.parser.parse_args()
@@ -60,6 +65,12 @@ class Item(Resource):
 
 
 class Items(Resource):
+    @jwt_optional
     def get(self):
-        return {'items': [item.json() for item in ItemModel.query.all()]}
+        user_id = get_jwt_identity()
+        items = [item.json() for item in ItemModel.query.all()]
+        if user_id:
+            return {'items': items}, 200
+        return {'items': [item['name'] for item in items],
+                'message': 'More data available if you log in '}, 200
